macro def for reusable table event handlers
